backend/bot/requirements.txt
# Mattermost Bot Dependencies
mattermostdriver>=7.3.2
requests>=2.28.2
python-dotenv>=1.0.0
